+++ auto_submit/lib/service/log.dart
@@ -1,0 +1,7 @@
+// Copyright 2019 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:logging/logging.dart';
+
+final Logger logger = Logger('auto_submit');
